ISLE: support more flexible integer constants. (#4559)
The ISLE language's lexer previously used a very primitive
`i64::from_str_radix` call to parse integer constants, allowing values
in the range -2^63..2^63 only. Also, underscores to separate digits (as
is allwoed in Rust) were not supported. Finally, 128-bit constants were
not supported at all.

This PR addresses all issues above:
- Integer constants are internally stored as 128-bit values.
- Parsing supports either signed (-2^127..2^127) or unsigned (0..2^128)
  range. Negation works independently of that, so one can write
  `-0xffff..ffff` (128 bits wide, i.e., -(2^128-1)) to get a `1`.
- Underscores are supported to separate groups of digits, so one can
  write `0xffff_ffff`.
- A minor oversight was fixed: hex constants can start with `0X`
  (uppercase) as well as `0x`, for consistency with Rust and C.

This PR also adds a new kind of ISLE test that actually runs a driver
linked to compiled ISLE code; we previously didn't have any such tests,
but it is now quite useful to assert correct interpretation of constant
values.

+++ cranelift/isle/isle/src/ast.rs
@@ -128,7 +128,7 @@ pub enum Pattern {
         pos: Pos,
     },
     /// An operator that matches a constant integer value.
-    ConstInt { val: i64, pos: Pos },
+    ConstInt { val: i128, pos: Pos },
     /// An operator that matches an external constant value.
     ConstPrim { val: Ident, pos: Pos },
     /// An application of a type variant or term.
@@ -313,7 +313,7 @@ pub enum Expr {
     /// A variable use.
     Var { name: Ident, pos: Pos },
     /// A constant integer.
-    ConstInt { val: i64, pos: Pos },
+    ConstInt { val: i128, pos: Pos },
     /// A constant of some other primitive type.
     ConstPrim { val: Ident, pos: Pos },
     /// The `(let ((var ty val)*) body)` form.
